requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==1.26.15
